third_party/blink/web_tests/virtual/fenced-frame-shadow-dom/wpt_internal/fenced_frame/storage-partitioning.https-expected.txt

This is a testharness.js-based test.
FAIL document.cookie assert_equals: The inner value is gone after an embedder-initiated refresh. expected (object) null but got (string) "inner"
FAIL localStorage assert_equals: The inner value is gone after an embedder-initiated refresh. expected (object) null but got (string) "inner"
Harness: the test ran to completion.
